adjust imports, move flags location
